requirements.txt
bayesian
pillow
-e git+https://github.com/django/django.git@stable/1.7.x#egg=django17
djangorestframework
jsonfield
djangorestframework-gis
